chore(cleanup): cleanup queries
